requirements.txt
alembic==1.4.2
blinker==1.4
certifi==2020.6.20
coverage==5.2.1
distlib==0.3.1
dnspython==1.16.0
email-validator==1.1.1
elasticsearch==7.9.1
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Moment==0.10.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
Pillow==7.2.0
pytest==6.0.1
python-dateutil==2.8.1
python-editor==1.0.4
python-engineio==3.13.2
pytz==2020.1
pyxdg==0.26
redis==3.5.3
requests==2.24.0
rq==1.5.2
service-identity==18.1.0
six==1.15.0
SQLAlchemy==1.3.19
urllib3==1.25.10
webencodings==0.5.1
Werkzeug==1.0.1
WTForms==2.3.3
zope.interface==5.1.2
